Add Generate command
